# LUCENE-8968: Improve performance of WITHIN and DISJOINT queries for Shape queries