… lucene/sandbox/src/java/org/apache/lucene/document/XYShapeBoundingBoxQuery.java
… lucene/sandbox/src/java/org/apache/lucene/document/XYShapeLineQuery.java
… lucene/sandbox/src/java/org/apache/lucene/document/XYShapePolygonQuery.java