@iverase 2019-09-05
Thanks @jpountz , I agree we should not be keeping a state in the visitor. I updated the PR going back to two passes but using only one dense bitset. In the case where `values.getDocCount() == reader.maxDoc()` we only need one pass.